Cleanup, restructure a bit, add comments. Seems to work.
