+++ frontend/src/Shelterly.js
@@ -1,81 +1,46 @@
-import React, {Fragment, useContext, useEffect, useReducer} from "react";
-import axios from "axios";
-import {navigate, useRoutes, A} from "hookrouter";
+import React, {Fragment, useContext, useEffect} from "react";
+import {A, navigate, useRoutes} from "hookrouter";
 import routes from "./router";
 import PageNotFound from "./components/PageNotFound";
-import {CounterContext} from "./accounts/AccountsReducer";
+import {AuthContext} from "./accounts/AccountsReducer";
 import setAuthToken from "./accounts/setAuthToken";
+import {loadUser, logoutUser} from "./accounts/AccountsUtils";
+import logo from "./static/images/nvadg_logo.png"
+
+const header_style = {
+  textAlign: "center",
+};
 
 if (localStorage.getItem('token')) setAuthToken(localStorage.getItem('token'));
 function Shelterly() {
 
   // Initial login state.
-  const { state, dispatch } = useContext(CounterContext);
+  const { state, dispatch } = useContext(AuthContext);
 
   useEffect(() => {
-    if (localStorage.getItem('token')) {
-      loadUser();
+    // If we have a token but no user, attempt to authenticate them.
+    if (!state.user && localStorage.getItem('token')) {
+      loadUser({dispatch});
     }
-    console.log("token "+localStorage.getItem('token'));
   }, []);
 
-  // Redirect to login page if user is logged out.
-  console.log(state);
+  // Redirect to login page if no authenticated user object is present.
   if (!state.user) {
-    console.log('logged out');
     navigate('/login');
   }
   const routeResult = useRoutes(routes);
 
-  // Login
-  function loadUser() {
-    if (localStorage.getItem('token')) setAuthToken(localStorage.getItem('token'));
-
-    // DISPATCH USER_LOADING
-    dispatch({ type: 'USER_LOADING' });
-
-    let headers = {
-      "Content-Type": "application/json",
-    };
-
-    axios.get("http://localhost:8000/accounts/auth/user/", {
-      headers: headers
-    })
-    .then(function(results){
-      console.log(results.data);
-      dispatch({type: 'USER_LOADED', user: results.data });
-    })
-    .catch(e => {
-      console.log('error: '+e);
-      dispatch({type: "AUTHENTICATION_ERROR", data: e});
-    })
-  }
-
-  // Logout
-  function logoutUser() {
-    let headers = {
-      "Content-Type": "application/json",
-    };
-  
-    axios.post("http://localhost:8000/logout/", {}, {
-      headers: headers
-    })
-    .then(function() {
-      dispatch({ type: 'LOGOUT_SUCCESSFUL' });
-      navigate('/login');
-    })
-    .catch(e => {
-      console.log(e);
-    });
-  }
-
   return (
     <div>
+      <h1 style={header_style} className="col-12">
+        <A href="/"><img src={logo} alt=""/></A>
+      </h1>
+      <hr className="mt-0 mb-4"/>
       <Fragment>
         {routeResult || <PageNotFound />}
       </Fragment>
       <div style={{textAlign: "right"}}>
-      {state.isAuthenticated ? <button onClick={logoutUser}>logout </button> : ''}
+        {state.user ? <button className="btn btn-danger" onClick={() => logoutUser({dispatch})}>Logout {state.user.username}</button> : ''}
       </div>
     </div>
   );
